Explicitly define all types and put them in shared/types.ts
